refactor: 통화 화면 공통된 ui CallViewLayout으로 분리
